expose a `createBrowserRouter` and associated functions
For consistency with `useRoutes` and `Routes`, let's expose pre-wrapped
functions to create routers. It's a bit easier to use, and we'll see if
we need to expose the "registerRouter" capability in the future.
