+++ packages/rum-react/src/domain/reactRouterV6/index.ts
@@ -1,3 +1,3 @@
-export { registerRouter } from './registerRouter'
+export { createMemoryRouter, createHashRouter, createBrowserRouter } from './createRouter'
 export { useRoutes } from './useRoutes'
 export { Routes } from './routesComponent'
